+++ lib/standard_id.rb
@@ -1,5 +1,6 @@
 require "standard_id/version"
 require "standard_id/deprecator"
+require "standard_id/instrumentation"
 require "standard_id/current_attributes"
 require "standard_id/engine"
 require "standard_id/web_engine"